requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
pyod
tensorflow
openpyxl
fuzzywuzzy
plotly
python-pptx
BeautifulSoup4
python-pptx
python-dotenv==1.0.0
langchain==0.0.184
PyPDF2==3.0.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
streamlit_chat

# -e .
